+++ src/AppServices/Enforcement/EnforcementActionQuery/IActionViewDto.cs
@@ -1,5 +1,5 @@
 ﻿using AirWeb.AppServices.Staff.Dto;
-using AirWeb.Domain.EnforcementEntities.Actions;
+using AirWeb.Domain.EnforcementEntities.EnforcementActions;
 
 namespace AirWeb.AppServices.Enforcement.EnforcementActionQuery;
