+++ src/Domain/EnforcementEntities/EnforcementActions/LetterOfNoncompliance.cs
@@ -1,18 +1,18 @@
 ﻿using AirWeb.Domain.EnforcementEntities.CaseFiles;
 using AirWeb.Domain.Identity;
 
-namespace AirWeb.Domain.EnforcementEntities.Actions;
+namespace AirWeb.Domain.EnforcementEntities.EnforcementActions;
 
-public class InformationalLetter : EnforcementAction, IResponseRequested
+public class LetterOfNoncompliance : EnforcementAction, IResponseRequested
 {
     // Constructors
     [UsedImplicitly] // Used by ORM.
-    private InformationalLetter() { }
+    private LetterOfNoncompliance() { }
 
-    internal InformationalLetter(Guid id, CaseFile caseFile, ApplicationUser? user)
+    internal LetterOfNoncompliance(Guid id, CaseFile caseFile, ApplicationUser? user)
         : base(id, caseFile, user)
     {
-        ActionType = EnforcementActionType.InformationalLetter;
+        ActionType = EnforcementActionType.LetterOfNoncompliance;
     }
 
     public void RequestResponse() => ResponseRequested = true;
