+++ src/LocalRepository/Repositories/LocalEnforcementActionRepository.cs
@@ -1,5 +1,5 @@
 using AirWeb.Domain.EnforcementEntities;
-using AirWeb.Domain.EnforcementEntities.Actions;
+using AirWeb.Domain.EnforcementEntities.EnforcementActions;
 using AirWeb.TestData.Enforcement;
 
 namespace AirWeb.LocalRepository.Repositories;
